refactor: core-infra, core-integration port 패키지 구분
