+++ src/main/java/com/bob/admin/member/application/port/result/ManagementMemberActivity.java
@@ -1,8 +1,5 @@
 package com.bob.admin.member.application.port.result;
 
-import com.bob.admin.member.application.port.result.activity.ManagementMemberPost;
-import com.bob.admin.member.application.port.result.activity.ManagementMemberTrade;
-
 public record ManagementMemberActivity(ManagementMemberPost post, ManagementMemberTrade trade) {
 
 }
