+++ src/main/java/com/bob/core/member/application/MemberCommandService.java
@@ -34,15 +34,15 @@ import com.bob.core.member.application.dto.command.SocialLoginCommand;
 import com.bob.core.member.application.port.in.MemberModifier;
 import com.bob.core.member.application.port.in.MemberReader;
 import com.bob.core.member.application.port.in.MemberRegister;
-import com.bob.core.member.application.port.out.MailSender;
 import com.bob.core.member.application.port.out.MemberAreaPort;
-import com.bob.core.member.application.port.out.MemberCachePort;
 import com.bob.core.member.application.port.out.MemberInterestPort;
+import com.bob.core.member.application.port.out.infra.MailSender;
+import com.bob.core.member.application.port.out.infra.MemberCachePort;
 import com.bob.core.member.domain.Member;
 import com.bob.core.member.domain.MemberArea;
 import com.bob.core.member.domain.MemberInterest;
-import com.bob.core.member.domain.PasswordEncoder;
 import com.bob.core.member.domain.Status;
+import com.bob.core.member.domain.encoder.PasswordEncoder;
 import com.bob.core.member.domain.repository.MemberRepository;
 import com.bob.core.member.event.MemberDeactivatedEvent;
 import com.bob.core.member.event.MemberRecoveredEvent;
