+++ src/main/java/com/bob/infrastructure/mail/adapter/GoogleMailSender.java
@@ -12,7 +12,7 @@ import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-import com.bob.core.member.application.port.out.MailSender;
+import com.bob.core.member.application.port.out.infra.MailSender;
 
 @Slf4j
 @Service
