+++ src/main/java/com/bob/security/config/SecurityConfig.java
@@ -40,8 +40,8 @@ import com.bob.security.adapter.handler.OAuth2FailureHandler;
 import com.bob.security.adapter.handler.OAuth2SuccessHandler;
 import com.bob.security.application.MemberDetailsService;
 import com.bob.security.application.OAuth2Service;
-import com.bob.security.application.port.out.AuthCachePort;
-import com.bob.security.application.port.out.TokenManager;
+import com.bob.security.application.port.out.infra.AuthCachePort;
+import com.bob.security.application.port.out.infra.TokenManager;
 import com.bob.security.config.props.HeaderProperties;
 import com.bob.security.config.registry.OptionalRegistry;
 import com.bob.security.config.registry.PermitAllRegistry;
