+++ src/test/java/com/bob/core/member/adapter/api/MemberApiTest.java
@@ -30,10 +30,10 @@ import com.bob.core.member.adapter.api.request.IssuePasswordRequest;
 import com.bob.core.member.adapter.api.request.RecoverAccountRequest;
 import com.bob.core.member.adapter.api.request.SignupRequest;
 import com.bob.core.member.application.dto.result.MemberDetail;
-import com.bob.core.member.application.port.out.MemberCachePort;
+import com.bob.core.member.application.port.out.infra.MemberCachePort;
 import com.bob.core.member.domain.Member;
-import com.bob.core.member.domain.PasswordEncoder;
 import com.bob.core.member.domain.Role;
+import com.bob.core.member.domain.encoder.PasswordEncoder;
 import com.bob.core.member.domain.repository.MemberRepository;
 import com.bob.security.model.MemberDetails;
 import com.bob.support.annotation.BobApiTest;
