+++ src/test/java/com/bob/security/adapter/filter/TokenAuthorizationFilterTest.java
@@ -33,7 +33,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 import com.bob.global.exception.exceptions.ApplicationAuthenticationException;
 import com.bob.security.adapter.entrypoint.TokenAuthenticationEntryPoint;
-import com.bob.security.application.port.out.TokenManager;
+import com.bob.security.application.port.out.infra.TokenManager;
 import com.bob.security.config.registry.OptionalRegistry;
 import com.bob.security.config.registry.PermitAllRegistry;
 import com.bob.security.model.MemberDetails;
